renaming, and updates to docs
